+++ aquadoggo/src/materializer/tasks/dependency.rs
@@ -4,7 +4,7 @@ use log::{debug, trace};
 use p2panda_rs::document::traits::AsDocument;
 use p2panda_rs::document::DocumentViewId;
 use p2panda_rs::operation::OperationValue;
-use p2panda_rs::schema::SchemaId;
+use p2panda_rs::schema::{FieldType, SchemaId};
 use p2panda_rs::storage_provider::traits::DocumentStore;
 
 use crate::context::Context;
@@ -44,7 +44,7 @@ pub async fn dependency_task(context: Context, input: TaskInput) -> TaskResult<T
 
     let document = match document {
         Some(document) => {
-            debug!("Document retrieved from storage with view id: {}", view_id);
+            trace!("Document retrieved from storage with view id: {}", view_id);
             Ok(document)
         }
         // If no document with the view for the id passed into this task could be retrieved then
@@ -87,8 +87,12 @@ pub async fn dependency_task(context: Context, input: TaskInput) -> TaskResult<T
                     "Pinned relation field found referring to view id: {}",
                     pinned_relation.view_id()
                 );
-                next_tasks
-                    .push(get_relation_task(&context, pinned_relation.view_id().clone()).await?);
+
+                if let Some(task) =
+                    get_relation_task(&context, pinned_relation.view_id().clone()).await?
+                {
+                    next_tasks.push(task);
+                }
             }
             OperationValue::PinnedRelationList(pinned_relation_list) => {
                 // same as above...
@@ -100,8 +104,12 @@ pub async fn dependency_task(context: Context, input: TaskInput) -> TaskResult<T
                             "Pinned relation list field found containing view id: {}",
                             document_view_id
                         );
-                        next_tasks
-                            .push(get_relation_task(&context, document_view_id.clone()).await?);
+
+                        if let Some(task) =
+                            get_relation_task(&context, document_view_id.clone()).await?
+                        {
+                            next_tasks.push(task);
+                        }
                     }
                 }
             }
@@ -110,32 +118,30 @@ pub async fn dependency_task(context: Context, input: TaskInput) -> TaskResult<T
     }
 
     // Construct additional tasks if the task input matches certain system schemas and all
-    // dependencies have been reduced.
-    let all_dependencies_met = !next_tasks.iter().any(|task| task.is_some());
-    if all_dependencies_met {
-        // Helper that returns a schema task for the current task input.
-        let schema_task = || {
-            Some(Task::new(
-                "schema",
-                TaskInput::DocumentViewId(document_view.id().clone()),
-            ))
-        };
-
+    // "child" dependencies have been reduced
+    let child_dependencies_met = next_tasks.is_empty();
+    if child_dependencies_met {
         match document.schema_id() {
             // Start `schema` task when a schema (field) definition view is completed with
             // dependencies
-            SchemaId::SchemaDefinition(_) => next_tasks.push(schema_task()),
-            SchemaId::SchemaFieldDefinition(_) => next_tasks.push(schema_task()),
+            SchemaId::SchemaDefinition(_) | SchemaId::SchemaFieldDefinition(_) => {
+                next_tasks.push(Task::new(
+                    "schema",
+                    TaskInput::DocumentViewId(document_view.id().clone()),
+                ));
+            }
             _ => {}
         }
     }
 
-    debug!(
-        "Scheduling {} reduce tasks",
-        next_tasks.iter().filter(|t| t.is_some()).count()
-    );
+    // Now we check all the "parent" or "inverse" relations, that is _other_ documents pointing at
+    // the one we're currently looking at
+    let mut reverse_tasks = get_inverse_relation_tasks(&context, document.schema_id()).await?;
+    next_tasks.append(&mut reverse_tasks);
 
-    Ok(Some(next_tasks.into_iter().flatten().collect()))
+    debug!("Scheduling {} tasks", next_tasks.len());
+
+    Ok(Some(next_tasks))
 }
 
 /// Returns a _reduce_ task for a given document view only if that view does not yet exist in the
@@ -166,18 +172,79 @@ async fn get_relation_task(
     }
 }
 
+/// Returns _dependency_ tasks for every document which has a pinned relation or pinned relation
+/// list to a document view with the given schema id.
+async fn get_inverse_relation_tasks(
+    context: &Context,
+    schema_id: &SchemaId,
+) -> Result<Vec<Task<TaskInput>>, TaskError> {
+    let mut tasks = Vec::new();
+
+    // Find all "parent" schemas which have at least one relation field pointing at documents of
+    // the given schema id
+    let parent_schema_ids: Vec<SchemaId> = context
+        .schema_provider
+        .all()
+        .await
+        .iter()
+        .filter_map(|schema| {
+            let has_relation_to_schema =
+                schema
+                    .fields()
+                    .iter()
+                    .any(|(_, field_type)| match field_type {
+                        FieldType::PinnedRelation(relation_schema_id)
+                        | FieldType::PinnedRelationList(relation_schema_id) => {
+                            relation_schema_id == schema_id
+                        }
+                        _ => false,
+                    });
+
+            if has_relation_to_schema {
+                Some(schema.id().to_owned())
+            } else {
+                None
+            }
+        })
+        .collect();
+
+    // Find all documents which follow these "parent" schemas
+    for parent_schema_id in parent_schema_ids {
+        // @TODO: Use a more efficient SQL query here, this does too much
+        let parent_documents = context
+            .store
+            .get_documents_by_schema(&parent_schema_id)
+            .await
+            .map_err(|err| TaskError::Critical(err.to_string()))?;
+
+        for parent_document in parent_documents {
+            // Dispatch dependency tasks from the latest document view of each parent document. We
+            // _only_ materialise historical document views when at least one _latest_ document
+            // view points at it.
+            tasks.push(Task::new(
+                "dependency",
+                TaskInput::DocumentViewId(parent_document.view_id().to_owned()),
+            ));
+        }
+    }
+
+    Ok(tasks)
+}
+
 #[cfg(test)]
 mod tests {
     use p2panda_rs::document::traits::AsDocument;
     use p2panda_rs::document::{DocumentId, DocumentViewId};
+    use p2panda_rs::entry::decode::decode_entry;
     use p2panda_rs::entry::traits::AsEncodedEntry;
     use p2panda_rs::identity::KeyPair;
+    use p2panda_rs::operation::encode::encode_operation;
     use p2panda_rs::operation::{
-        Operation, OperationBuilder, OperationId, OperationValue, PinnedRelation,
+        Operation, OperationAction, OperationBuilder, OperationId, OperationValue, PinnedRelation,
         PinnedRelationList, Relation, RelationList,
     };
     use p2panda_rs::schema::{FieldType, Schema, SchemaId};
-    use p2panda_rs::storage_provider::traits::{DocumentStore, OperationStore};
+    use p2panda_rs::storage_provider::traits::{DocumentStore, EntryStore, OperationStore};
     use p2panda_rs::test_utils::fixtures::{key_pair, random_document_id, random_document_view_id};
     use p2panda_rs::test_utils::memory_store::helpers::{
         populate_store, send_to_store, PopulateStoreConfig,
@@ -186,10 +253,10 @@ mod tests {
     use rstest::rstest;
 
     use crate::materializer::tasks::reduce_task;
-    use crate::materializer::TaskInput;
+    use crate::materializer::{Task, TaskInput};
     use crate::test_utils::{
         add_document, add_schema, doggo_schema, populate_store_config, schema_from_fields,
-        test_runner, TestNode,
+        test_runner, test_runner_with_manager, TestNode, TestNodeManager,
     };
 
     use super::dependency_task;
@@ -613,7 +680,8 @@ mod tests {
         config: PopulateStoreConfig,
     ) {
         test_runner(move |node: TestNode| async move {
-            // Populate the store with some entries and operations but DON'T materialise any resulting documents.
+            // Populate the store with some entries and operations but DON'T materialise any
+            // resulting documents.
             let (_, document_ids) = populate_store(&node.context.store, &config).await;
             let schema_field_document_id = document_ids
                 .get(0)
@@ -652,6 +720,378 @@ mod tests {
 
             let schema_tasks = tasks.iter().filter(|t| t.worker_name() == "schema").count();
             assert_eq!(schema_tasks, expected_schema_tasks);
+        });
+    }
+
+    #[rstest]
+    fn materialise_late_views(key_pair: KeyPair) {
+        test_runner_with_manager(|manager: TestNodeManager| async move {
+            let mut node_a = manager.create().await;
+            let node_b = manager.create().await;
+
+            // PREPARE TEST DATA
+
+            // Create a post schema with comments pinned to a particular version of each post
+            let post_schema = add_schema(
+                &mut node_a,
+                "post",
+                vec![("title", FieldType::String)],
+                &key_pair,
+            )
+            .await;
+
+            let comment_schema = add_schema(
+                &mut node_a,
+                "comment",
+                vec![
+                    ("post", FieldType::PinnedRelation(post_schema.id().clone())),
+                    ("text", FieldType::String),
+                ],
+                &key_pair,
+            )
+            .await;
+
+            // 1. Create document with cooking recipe
+            let post_operation_1 = OperationBuilder::new(post_schema.id())
+                .fields(&[("title", "How to make potato soup".into())])
+                .build()
+                .unwrap();
+
+            let (post_entry_1, _) = send_to_store(
+                &node_a.context.store,
+                &post_operation_1,
+                &post_schema,
+                &key_pair,
+            )
+            .await
+            .unwrap();
+
+            let post_operation_id_1 = post_entry_1.hash().into();
+            let post_document_id = DocumentId::new(&post_operation_id_1);
+            let post_view_id_1 = DocumentViewId::new(&[post_operation_id_1.clone()]);
+
+            // 2. Update it (later this will be the version someone will write a comment about)
+            let post_operation_2 = OperationBuilder::new(post_schema.id())
+                .action(OperationAction::Update)
+                .previous(&post_view_id_1)
+                .fields(&[(
+                    "title",
+                    "How to make potato soup, definitive edition!".into(),
+                )])
+                .build()
+                .unwrap();
+
+            let (post_entry_2, _) = send_to_store(
+                &node_a.context.store,
+                &post_operation_2,
+                &post_schema,
+                &key_pair,
+            )
+            .await
+            .unwrap();
+
+            let post_operation_id_2: OperationId = post_entry_2.hash().into();
+            let post_view_id_2 = DocumentViewId::new(&[post_operation_id_2.clone()]);
+
+            // 3. Update it, so there is another version after the one someone was commenting on
+            let post_operation_3 = OperationBuilder::new(post_schema.id())
+                .action(OperationAction::Update)
+                .previous(&post_view_id_2)
+                .fields(&[(
+                    "title",
+                    "How to make the ULTIMATE potato soup! Now for real, with bamboo grass.".into(),
+                )])
+                .build()
+                .unwrap();
+
+            let (post_entry_3, _) = send_to_store(
+                &node_a.context.store,
+                &post_operation_3,
+                &post_schema,
+                &key_pair,
+            )
+            .await
+            .unwrap();
+
+            let post_operation_id_3: OperationId = post_entry_3.hash().into();
+            let post_view_id_3 = DocumentViewId::new(&[post_operation_id_3.clone()]);
+
+            // 4. Write the comment on the recipe post, when it was in its second version
+            let comment_operation_1 = OperationBuilder::new(comment_schema.id())
+                .fields(&[(
+                    "text",
+                    "Sorry, but I think my potato soup is much better. Did you try bamboo grass?"
+                        .into(),
+                ), (
+                    "post",
+                    post_view_id_2.clone().into(),
+                )])
+                .build()
+                .unwrap();
+
+            let (comment_entry_1, _) = send_to_store(
+                &node_a.context.store,
+                &comment_operation_1,
+                &comment_schema,
+                &key_pair,
+            )
+            .await
+            .unwrap();
+
+            let comment_operation_id_1 = comment_entry_1.hash().into();
+            let comment_document_id = DocumentId::new(&comment_operation_id_1);
+            let comment_view_id_1 = DocumentViewId::new(&[comment_operation_id_1.clone()]);
+
+            // INSERT SCHEMAS INTO NODE B
+
+            node_b.context.schema_provider.update(post_schema).await;
+            node_b.context.schema_provider.update(comment_schema).await;
+
+            // INSERT "COMMENT" INTO NODE B'S DATABASE
+
+            // We do this to enforce an ordering of when the system processes what information. In
+            // reality that ordering might look different / be random.
+
+            node_b
+                .context
+                .store
+                .insert_operation(
+                    &comment_operation_id_1,
+                    &key_pair.public_key(),
+                    &comment_operation_1,
+                    &comment_document_id,
+                )
+                .await
+                .unwrap();
+
+            node_b
+                .context
+                .store
+                .insert_entry(
+                    &decode_entry(&comment_entry_1).unwrap(),
+                    &comment_entry_1,
+                    Some(&encode_operation(&comment_operation_1).unwrap()),
+                )
+                .await
+                .unwrap();
+
+            // 1. Materialise the "comment" document first, it will point at a post which does not
+            //    exist yet ..
+            let input = TaskInput::DocumentId(comment_document_id);
+            let tasks = reduce_task(node_b.context.clone(), input)
+                .await
+                .unwrap()
+                .expect("Should have returned new tasks");
+            assert_eq!(tasks.len(), 1);
+            assert_eq!(tasks[0].worker_name(), &String::from("dependency"));
+
+            // 2. The "dependency" task will try to resolve the pinned document view pointing at
+            //    the "post" document in it's version 2
+            let tasks = dependency_task(node_b.context.clone(), tasks[0].input().clone())
+                .await
+                .unwrap();
+            assert_eq!(
+                tasks,
+                Some(vec![Task::new(
+                    "reduce",
+                    TaskInput::DocumentViewId(post_view_id_2.clone())
+                )])
+            );
+
+            // 3. .. this kicks in a "reduce" task for that particular view, but it will cancel
+            //    since not enough data exists to materialize the "post" document
+            let tasks = reduce_task(
+                node_b.context.clone(),
+                TaskInput::DocumentViewId(post_view_id_2.clone()),
+            )
+            .await
+            .unwrap();
+            assert!(tasks.is_none());
+
+            // INSERT "POST" INTO NODE B'S DATABASE
+
+            node_b
+                .context
+                .store
+                .insert_operation(
+                    &post_operation_id_1,
+                    &key_pair.public_key(),
+                    &post_operation_1,
+                    &post_document_id,
+                )
+                .await
+                .unwrap();
+
+            node_b
+                .context
+                .store
+                .insert_entry(
+                    &decode_entry(&post_entry_1).unwrap(),
+                    &post_entry_1,
+                    Some(&encode_operation(&post_operation_1).unwrap()),
+                )
+                .await
+                .unwrap();
+
+            node_b
+                .context
+                .store
+                .insert_operation(
+                    &post_operation_id_2,
+                    &key_pair.public_key(),
+                    &post_operation_2,
+                    &post_document_id,
+                )
+                .await
+                .unwrap();
+
+            node_b
+                .context
+                .store
+                .insert_entry(
+                    &decode_entry(&post_entry_2).unwrap(),
+                    &post_entry_2,
+                    Some(&encode_operation(&post_operation_2).unwrap()),
+                )
+                .await
+                .unwrap();
+
+            node_b
+                .context
+                .store
+                .insert_operation(
+                    &post_operation_id_3,
+                    &key_pair.public_key(),
+                    &post_operation_3,
+                    &post_document_id,
+                )
+                .await
+                .unwrap();
+
+            node_b
+                .context
+                .store
+                .insert_entry(
+                    &decode_entry(&post_entry_3).unwrap(),
+                    &post_entry_3,
+                    Some(&encode_operation(&post_operation_3).unwrap()),
+                )
+                .await
+                .unwrap();
+
+            // 1. Now materialise the post, it will exist now in its latest form, but the comment
+            //    will point at an old version of it, and this historical view does not exist yet
+            //    ..
+            let input = TaskInput::DocumentId(post_document_id);
+            let tasks = reduce_task(node_b.context.clone(), input)
+                .await
+                .unwrap()
+                .expect("Should have returned new tasks");
+            assert_eq!(tasks.len(), 1);
+            assert_eq!(tasks[0].worker_name(), &String::from("dependency"));
+
+            // We should have now a materialized latest post and comment document but not the
+            // pinned historical version of the post, where the comment was pointing at!
+            assert!(node_b
+                .context
+                .store
+                .get_document_by_view_id(&post_view_id_3)
+                .await
+                .unwrap()
+                .is_some());
+
+            assert!(node_b
+                .context
+                .store
+                .get_document_by_view_id(&comment_view_id_1)
+                .await
+                .unwrap()
+                .is_some());
+
+            assert!(node_b
+                .context
+                .store
+                .get_document_by_view_id(&post_view_id_2)
+                .await
+                .unwrap()
+                .is_none()); // Should not exist yet!
+
+            // 2. The "dependency" task followed materialising the "post" found a reverse relation
+            //    to a "comment" document .. it dispatches another "dependency" task for it
+            let tasks = dependency_task(node_b.context.clone(), tasks[0].input().clone())
+                .await
+                .unwrap();
+            assert_eq!(
+                tasks,
+                Some(vec![Task::new(
+                    "dependency",
+                    TaskInput::DocumentViewId(comment_view_id_1.clone())
+                )])
+            );
+
+            // 3. Running the "dependency" task again over the "comment" document we finally get
+            //    back to the missing pinned relation to the older version of the "post" document
+            let tasks = dependency_task(
+                node_b.context.clone(),
+                TaskInput::DocumentViewId(comment_view_id_1.clone()),
+            )
+            .await
+            .unwrap();
+            assert_eq!(
+                tasks,
+                Some(vec![Task::new(
+                    "reduce",
+                    TaskInput::DocumentViewId(post_view_id_2.clone())
+                )])
+            );
+
+            // 4. We finally materialise the missing view!
+            let tasks = reduce_task(
+                node_b.context.clone(),
+                TaskInput::DocumentViewId(post_view_id_2.clone()),
+            )
+            .await
+            .unwrap();
+            assert_eq!(
+                tasks,
+                Some(vec![Task::new(
+                    "dependency",
+                    TaskInput::DocumentViewId(post_view_id_2.clone())
+                )])
+            );
+
+            // It exists now! :-)
+            assert!(node_b
+                .context
+                .store
+                .get_document_by_view_id(&post_view_id_2)
+                .await
+                .unwrap()
+                .is_some());
+
+            // 5. Running the two dispatched "dependency" tasks again ends with no more tasks. This
+            //    is good, otherwise we would run into an infinite, recursive loop
+            let tasks = dependency_task(
+                node_b.context.clone(),
+                TaskInput::DocumentViewId(post_view_id_2.clone()),
+            )
+            .await
+            .unwrap();
+            assert_eq!(
+                tasks,
+                Some(vec![Task::new(
+                    "dependency",
+                    TaskInput::DocumentViewId(comment_view_id_1.clone())
+                )])
+            );
+
+            let tasks = dependency_task(
+                node_b.context.clone(),
+                TaskInput::DocumentViewId(comment_view_id_1.clone()),
+            )
+            .await
+            .unwrap();
+            assert_eq!(tasks, Some(vec![]));
         });
     }
 }