+++ aquadoggo/src/materializer/tasks/reduce.rs
@@ -2,7 +2,7 @@
 
 use std::convert::TryFrom;
 
-use log::{debug, info};
+use log::{debug, info, trace};
 use p2panda_rs::document::traits::AsDocument;
 use p2panda_rs::document::{Document, DocumentBuilder, DocumentId, DocumentViewId};
 use p2panda_rs::operation::traits::{AsOperation, WithPublicKey};
@@ -68,10 +68,10 @@ async fn resolve_document_id<S: EntryStore + OperationStore + LogStore + Documen
         }
         TaskInput::DocumentViewId(document_view_id) => {
             // Document view id is given, let's find out its document id
-            debug!("Find document for view with id: {}", document_view_id);
+            trace!("Find document for view with id: {}", document_view_id);
 
             // Pick one operation from the view, this is all we need to determine the document id
-            let operation_id = document_view_id.iter().next().unwrap();
+            let operation_id = document_view_id.iter().last().unwrap();
 
             // Determine document id by looking into the operations stored on the node already.
             // Note, finding a document id here does not mean the document has been materialized
@@ -97,6 +97,41 @@ async fn reduce_document_view<O: AsOperation + WithId<OperationId> + WithPublicK
     document_view_id: &DocumentViewId,
     operations: &Vec<O>,
 ) -> Result<Option<Vec<Task<TaskInput>>>, TaskError> {
+    // Attempt to retrieve the document this view is part of in order to determine if it has
+    // been once already materialized yet.
+    // @TODO: This can be a more efficient storage method. See issue:
+    // https://github.com/p2panda/aquadoggo/issues/431
+    let document_exists = context
+        .store
+        .get_document(document_id)
+        .await
+        .map_err(|err| TaskError::Critical(err.to_string()))?
+        .is_some();
+
+    // If it wasn't found then we shouldn't reduce this view yet (as the document it's part of
+    // should be reduced first).
+    //
+    // In this case, we assume the task for reducing the document is triggered in the future which
+    // is followed by a dependency task finally looking at all view ids of this document (through
+    // pinned relation ids pointing at it).
+    if !document_exists {
+        return Ok(None);
+    };
+
+    // Make sure to not materialize and store document view twice
+    // @TODO: This can be a more efficient storage method. See issue:
+    // https://github.com/p2panda/aquadoggo/issues/431
+    let document_view_exists = context
+        .store
+        .get_document_by_view_id(document_view_id)
+        .await
+        .map_err(|err| TaskError::Critical(err.to_string()))?
+        .is_some();
+
+    if document_view_exists {
+        return Ok(None);
+    }
+
     // Materialize document view
     let document_builder: DocumentBuilder = operations.into();
     let document = match document_builder.build_to_view_id(Some(document_view_id.to_owned())) {
@@ -125,62 +160,27 @@ async fn reduce_document_view<O: AsOperation + WithId<OperationId> + WithPublicK
         }
     };
 
-    // Attempt to retrieve the document this view is part of in order to determine if it has
-    // been once already materialized yet.
-    let existing_document = context
+    context
         .store
-        .get_document(document_id)
+        .insert_document_view(
+            &document.view().unwrap(),
+            document.id(),
+            document.schema_id(),
+        )
         .await
         .map_err(|err| TaskError::Critical(err.to_string()))?;
 
-    // If it wasn't found then we shouldn't reduce this view yet (as the document it's part of
-    // should be reduced first). In this case we assume the task for reducing the document is
-    // already in the queue and so we simply re-issue this reduce task.
-    if existing_document.is_none() {
-        debug!(
-            "Document {} for view not materialized yet, reissuing current reduce task for {}",
-            document_id.display(),
-            document_view_id.display()
-        );
-        return Ok(Some(vec![Task::new(
-            "reduce",
-            TaskInput::DocumentViewId(document_view_id.to_owned()),
-        )]));
-    };
+    info!("Stored {} document view {}", document, document.view_id());
 
-    // Make sure to not store document view twice
-    let document_view_exists = context
-        .store
-        .get_document_by_view_id(document_view_id)
-        .await
-        .map_err(|err| TaskError::Critical(err.to_string()))?
-        .is_some();
+    debug!(
+        "Dispatch dependency task for view with id: {}",
+        document.view_id()
+    );
 
-    if !document_view_exists {
-        context
-            .store
-            .insert_document_view(
-                &document.view().unwrap(),
-                document.id(),
-                document.schema_id(),
-            )
-            .await
-            .map_err(|err| TaskError::Critical(err.to_string()))?;
-
-        info!("Stored {} document view {}", document, document.view_id());
-
-        debug!(
-            "Dispatch dependency task for view with id: {}",
-            document.view_id()
-        );
-
-        Ok(Some(vec![Task::new(
-            "dependency",
-            TaskInput::DocumentViewId(document.view_id().to_owned()),
-        )]))
-    } else {
-        Ok(None)
-    }
+    Ok(Some(vec![Task::new(
+        "dependency",
+        TaskInput::DocumentViewId(document.view_id().to_owned()),
+    )]))
 }
 
 /// Helper method to reduce an operation graph to the latest document view, returning the
@@ -194,6 +194,20 @@ async fn reduce_document<O: AsOperation + WithId<OperationId> + WithPublicKey>(
 ) -> Result<Option<Vec<Task<TaskInput>>>, TaskError> {
     match Document::try_from(operations) {
         Ok(document) => {
+            // Make sure to not materialize and store document view twice
+            // @TODO: This can be a more efficient storage method. See issue:
+            // https://github.com/p2panda/aquadoggo/issues/431
+            let document_view_exists = context
+                .store
+                .get_document_by_view_id(document.view_id())
+                .await
+                .map_err(|err| TaskError::Critical(err.to_string()))?
+                .is_some();
+
+            if document_view_exists {
+                return Ok(None);
+            }
+
             // Insert this document into storage. If it already existed, this will update it's
             // current view
             context
@@ -219,7 +233,7 @@ async fn reduce_document<O: AsOperation + WithId<OperationId> + WithPublicKey>(
                     document.view_id().display()
                 );
             } else {
-                info!("Created {}", document.display(),);
+                info!("Created {}", document.display());
             };
 
             debug!(
@@ -260,7 +274,7 @@ mod tests {
     use rstest::rstest;
 
     use crate::materializer::tasks::reduce_task;
-    use crate::materializer::{Task, TaskInput};
+    use crate::materializer::TaskInput;
     use crate::test_utils::{
         doggo_fields, doggo_schema, populate_store_config, test_runner, TestNode,
     };
@@ -450,53 +464,6 @@ mod tests {
                 document.unwrap().get("username").unwrap(),
                 &OperationValue::String("bubu".to_string())
             );
-        });
-    }
-
-    #[rstest]
-    fn reissue_reduce_when_document_does_not_exist(
-        #[from(populate_store_config)]
-        #[with(
-            2,
-            1,
-            1,
-            false,
-            doggo_schema(),
-            doggo_fields(),
-            vec![("username", OperationValue::String("PANDA".into()))]
-        )]
-        config: PopulateStoreConfig,
-    ) {
-        test_runner(move |node: TestNode| async move {
-            // Populate the store with some entries and operations but DON'T materialise any
-            // resulting documents
-            let (_, document_ids) = populate_store(&node.context.store, &config).await;
-            let document_id = document_ids
-                .get(0)
-                .expect("There should be at least one document id");
-
-            // Get the operations
-            let document_operations = node
-                .context
-                .store
-                .get_operations_by_document_id(document_id)
-                .await
-                .unwrap();
-
-            // Build the document.
-            let document = DocumentBuilder::from(&document_operations).build().unwrap();
-
-            // Run a reduce task for this documents current view. The operations are already in
-            // the store, but the document is not materialized yet. This means we shouldn't insert
-            // any views for it yet. In that case, we expect this task to issue another reduce
-            // task with the same input.
-            let input = TaskInput::DocumentViewId(document.view_id().clone());
-            let next_tasks = reduce_task(node.context.clone(), input.clone())
-                .await
-                .expect("Task should succeed")
-                .expect("Task to be returned");
-
-            assert_eq!(next_tasks, vec![Task::new("reduce", input)]);
         });
     }